modificacion correcta de controlador y handler en funciones
